Move DetailsTabLiteratures to elements subfolder
