+++ app/packs/src/components/elements/DetailsTabLiteratures.js
